requirements.txt
gunicorn>=20.1.0
numpy>=1.26.0
pandas>=2.0.1
scikit-learn>=1.2.2
flask>=2.3.2
flask-wtf>=1.1.1
psycopg2-binary>=2.9.6
plotly>=5.14.1
sqlalchemy>=2.0.15
python-dotenv>=1.0.0
setuptools>=65.5.0
wheel>=0.38.0
